requirements.txt
pip
boto3==1.11.14
celery==4.3.0
colour==0.1.5
dj-database-url==0.5.0
django-axes==5.4.1
django-countries==5.5
django-db-geventpool==3.1.0
django-dirtyfields
django-environ==0.4.5
django-extensions==2.2.3
django_pglocks==1.0.3
django-redis==4.10.0
django-silk==3.0.1
django-storages==1.9.1
django-timezone-field==3.1
Django==2.2.6
djangorestframework==3.10.3
dpath==1.4.2
elasticsearch-dsl>=6.0.0,<7.0.0
feed-gov-back==0.4
freezegun==0.3.11
geckoboard.py==1.0.0
gevent==1.4.0
graphviz==0.8.3
gunicorn==20.0.4
ipython==7.8.0
minio==4.0.11
Markdown==3.1.1
notifications-python-client==5.3.0
openpyxl==2.5.12
odfpy==1.4.1
phonenumbers==8.10.20
psycogreen==1.0.1
psycopg2-binary==2.8.3
python-docx==0.8.10
python-pptx==0.6.18
PyPDF2==1.26.0
pytz
sentry-sdk==0.13.1
striprtf==0.0.10
redis==3.3.8
requests==2.22.0
requests_toolbelt==0.9.1
titlecase==0.12.0
werkzeug==0.14.1
whitenoise==3.3.1
xlrd==1.2.0
